+++ plom/client/chooser.py
@@ -15,6 +15,7 @@ __license__ = "AGPL-3.0-or-later"
 from datetime import datetime
 import logging
 from pathlib import Path
+import tempfile
 
 import toml
 import appdirs
@@ -28,6 +29,7 @@ from plom import Plom_API_Version
 from plom import Default_Port
 from plom.plom_exceptions import *
 from plom.messenger import Messenger
+from plom.client.comment_list import comment_file
 
 from .uiFiles.ui_chooser import Ui_Chooser
 from .useful_classes import ErrorMessage, SimpleMessage, ClientSettingsDialog
@@ -46,6 +48,7 @@ lastTime = {}
 log = logging.getLogger("client")
 logdir = Path(appdirs.user_log_dir("plom", "PlomGrading.org"))
 cfgdir = Path(appdirs.user_config_dir("plom", "PlomGrading.org"))
+cfgfilename = "plomConfig.toml"
 
 
 def readLastTime():
@@ -66,7 +69,7 @@ def readLastTime():
     lastTime["MarkWarnings"] = True
     # If config file exists, use it to update the defaults
     # prefer a local file, else try standard config location
-    for cfg in (Path("plomConfig.toml"), cfgdir / "plomConfig.toml"):
+    for cfg in (Path(cfgfilename), cfgdir / cfgfilename):
         if cfg.exists():
             # too early to log: log.info("Loading config file %s", cfg)
             with open(cfg) as data_file:
@@ -76,7 +79,7 @@ def readLastTime():
 
 def writeLastTime():
     """Write the options to the config file."""
-    cfg = cfgdir / "plomConfig.toml"
+    cfg = cfgdir / cfgfilename
     log.info("Saving config file %s", cfg)
     try:
         cfg.parent.mkdir(exist_ok=True)
@@ -162,7 +165,9 @@ class Chooser(QDialog):
         self.ui.mportSB.setValue(int(p))
 
     def options(self):
-        d = ClientSettingsDialog(lastTime)
+        d = ClientSettingsDialog(
+            lastTime, logdir, cfgdir / cfgfilename, tempfile.gettempdir(), comment_file
+        )
         d.exec_()
         # TODO: do something more proper like QSettings
         stuff = d.getStuff()